Fix for cross zone mark npc when entity ids are reused.

+++ common/repositories/base/base_raid_details_repository.h
@@ -16,6 +16,7 @@
 #include "../../strings.h"
 #include <ctime>
 
+
 class BaseRaidDetailsRepository {
 public:
 	struct RaidDetails {
@@ -23,9 +24,9 @@ public:
 		int32_t     loottype;
 		int8_t      locked;
 		std::string motd;
-		uint16_t    marked_npc_1;
-		uint16_t    marked_npc_2;
-		uint16_t    marked_npc_3;
+		uint32_t    marked_npc_1;
+		uint32_t    marked_npc_2;
+		uint32_t    marked_npc_3;
 	};
 
 	static std::string PrimaryKey()
@@ -96,10 +97,10 @@ public:
 	{
 		RaidDetails e{};
 
-		e.raidid      = 0;
-		e.loottype    = 0;
-		e.locked      = 0;
-		e.motd        = "";
+		e.raidid       = 0;
+		e.loottype     = 0;
+		e.locked       = 0;
+		e.motd         = "";
 		e.marked_npc_1 = 0;
 		e.marked_npc_2 = 0;
 		e.marked_npc_3 = 0;
@@ -139,13 +140,13 @@ public:
 		if (results.RowCount() == 1) {
 			RaidDetails e{};
 
-			e.raidid      = static_cast<int32_t>(atoi(row[0]));
-			e.loottype    = static_cast<int32_t>(atoi(row[1]));
-			e.locked      = static_cast<int8_t>(atoi(row[2]));
-			e.motd        = row[3] ? row[3] : "";
-			e.marked_npc_1 = static_cast<uint16_t>(strtoul(row[4], nullptr, 10));
-			e.marked_npc_2 = static_cast<uint16_t>(strtoul(row[5], nullptr, 10));
-			e.marked_npc_3 = static_cast<uint16_t>(strtoul(row[6], nullptr, 10));
+			e.raidid       = static_cast<int32_t>(atoi(row[0]));
+			e.loottype     = static_cast<int32_t>(atoi(row[1]));
+			e.locked       = static_cast<int8_t>(atoi(row[2]));
+			e.motd         = row[3] ? row[3] : "";
+			e.marked_npc_1 = static_cast<uint32_t>(strtoul(row[4], nullptr, 10));
+			e.marked_npc_2 = static_cast<uint32_t>(strtoul(row[5], nullptr, 10));
+			e.marked_npc_3 = static_cast<uint32_t>(strtoul(row[6], nullptr, 10));
 
 			return e;
 		}
@@ -283,13 +284,13 @@ public:
 		for (auto row = results.begin(); row != results.end(); ++row) {
 			RaidDetails e{};
 
-			e.raidid      = static_cast<int32_t>(atoi(row[0]));
-			e.loottype    = static_cast<int32_t>(atoi(row[1]));
-			e.locked      = static_cast<int8_t>(atoi(row[2]));
-			e.motd        = row[3] ? row[3] : "";
-			e.marked_npc_1 = static_cast<uint16_t>(strtoul(row[4], nullptr, 10));
-			e.marked_npc_2 = static_cast<uint16_t>(strtoul(row[5], nullptr, 10));
-			e.marked_npc_3 = static_cast<uint16_t>(strtoul(row[6], nullptr, 10));
+			e.raidid       = static_cast<int32_t>(atoi(row[0]));
+			e.loottype     = static_cast<int32_t>(atoi(row[1]));
+			e.locked       = static_cast<int8_t>(atoi(row[2]));
+			e.motd         = row[3] ? row[3] : "";
+			e.marked_npc_1 = static_cast<uint32_t>(strtoul(row[4], nullptr, 10));
+			e.marked_npc_2 = static_cast<uint32_t>(strtoul(row[5], nullptr, 10));
+			e.marked_npc_3 = static_cast<uint32_t>(strtoul(row[6], nullptr, 10));
 
 			all_entries.push_back(e);
 		}
@@ -314,13 +315,13 @@ public:
 		for (auto row = results.begin(); row != results.end(); ++row) {
 			RaidDetails e{};
 
-			e.raidid      = static_cast<int32_t>(atoi(row[0]));
-			e.loottype    = static_cast<int32_t>(atoi(row[1]));
-			e.locked      = static_cast<int8_t>(atoi(row[2]));
-			e.motd        = row[3] ? row[3] : "";
-			e.marked_npc_1 = static_cast<uint16_t>(strtoul(row[4], nullptr, 10));
-			e.marked_npc_2 = static_cast<uint16_t>(strtoul(row[5], nullptr, 10));
-			e.marked_npc_3 = static_cast<uint16_t>(strtoul(row[6], nullptr, 10));
+			e.raidid       = static_cast<int32_t>(atoi(row[0]));
+			e.loottype     = static_cast<int32_t>(atoi(row[1]));
+			e.locked       = static_cast<int8_t>(atoi(row[2]));
+			e.motd         = row[3] ? row[3] : "";
+			e.marked_npc_1 = static_cast<uint32_t>(strtoul(row[4], nullptr, 10));
+			e.marked_npc_2 = static_cast<uint32_t>(strtoul(row[5], nullptr, 10));
+			e.marked_npc_3 = static_cast<uint32_t>(strtoul(row[6], nullptr, 10));
 
 			all_entries.push_back(e);
 		}
